backend/requirements.txt
Flask
flask_cors
requests
python-dotenv
gunicorn
prometheus_client